fix: resolve all compilation errors for AyuGram features

+++ TMessagesProj/src/main/java/com/radolyn/ayugram/proprietary/AyuHistoryHook.java
@@ -45,7 +45,7 @@ public abstract class AyuHistoryHook {
             return;
         }
 
-        ArrayList<Integer> usersToLoad = new ArrayList<>();
+        ArrayList<Long> usersToLoad = new ArrayList<>();
         ArrayList<Long> chatsToLoad = new ArrayList<>();
 
         for (DeletedMessageFull deletedMessageFull : deletedMessages) {
